pyproject.toml
[project]
name = "wallpaper-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "pyobjc>=11.0",
]
